% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{convert_to_plotly}
\alias{convert_to_plotly}
\title{Convert CPIA Plot to Interactive Plotly}
\usage{
convert_to_plotly(plot)
}
\arguments{
\item{plot}{A styled ggplot object from style_cpia_plot()}
}
\value{
A plotly htmlwidget object with interactive features enabled.
The custom "text" aesthetic from ggplot is used for hover tooltips.
}
\description{
Converts a static ggplot object to an interactive plotly visualization. The
conversion enables hover tooltips, zooming, panning, and other interactive features.
Hover mode is set to "closest" so users can easily inspect individual data points.
}
\examples{
\dontrun{
styled_plot <- style_cpia_plot(base_plot, "q12b")
interactive_plot <- convert_to_plotly(styled_plot)
}

}
\keyword{internal}
