% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{create_question_info_modal}
\alias{create_question_info_modal}
\title{Create Question Info Modal}
\usage{
create_question_info_modal(
  question_code,
  question_label,
  indicator_info,
  sources,
  n_indicators,
  documentation_url = "https://rpubs.com/ifeanyi588/cpiascoringmethod"
)
}
\arguments{
\item{question_code}{The question code (e.g., "q12a")}

\item{question_label}{The full question label}

\item{indicator_info}{Tibble with columns: indicator, var_description_short, source}

\item{sources}{Character vector of data source names}

\item{n_indicators}{Number of indicators}

\item{documentation_url}{URL to full documentation (default: RPubs link)}
}
\value{
A shiny modal dialog UI element
}
\description{
Creates a modal dialog displaying data sources and indicators for a selected
CPIA question. Shows the actual indicators from cpiaetl::metadata_cpia that
contribute to the question's score with their descriptions. Links to full
documentation on RPubs.
}
\examples{
\dontrun{
# In module server
observeEvent(input$question_info, {
  sources <- get_question_data_sources()
  info <- sources[sources$variable == input$question, ]
  showModal(create_question_info_modal(
    question_code = input$question,
    question_label = "Property Rights",
    indicator_info = info$indicator_info[[1]],
    sources = info$sources[[1]],
    n_indicators = info$n_indicators
  ))
})
}

}
\keyword{internal}
