% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cpiaapp.R
\name{run_cpiaapp}
\alias{run_cpiaapp}
\title{Run the CPIA Shiny Application}
\usage{
run_cpiaapp(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A Shiny app object.
}
\description{
Launches an interactive Shiny dashboard for CPIA data visualization and analysis.
}
\examples{
\dontrun{
run_cpiaapp()
}

}
