% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{validate_datasets}
\alias{validate_datasets}
\title{Validate CPIA Datasets}
\usage{
validate_datasets(
  standard_data,
  africaii_data,
  group_standard_data,
  group_africaii_data
)
}
\arguments{
\item{standard_data}{Standard CPIA dataset}

\item{africaii_data}{African Integrity Indicators CPIA dataset}

\item{group_standard_data}{Standard group averages dataset}

\item{group_africaii_data}{African Integrity Indicators group averages dataset}
}
\value{
NULL (invisible). Function is called for its side effects (error throwing).
}
\description{
Validates that all required datasets are present and are valid data frames.
Throws an error if any dataset is NULL or not a data frame.
}
\examples{
\dontrun{
validate_datasets(
  standard_data = cpiaetl::standard_cpia,
  africaii_data = cpiaetl::africaii_cpia,
  group_standard_data = cpiaetl::group_standard_cpia,
  group_africaii_data = cpiaetl::group_africaii_cpia
)
}

}
\keyword{internal}
