% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module.R, R/viz_server.R
\name{viz_server}
\alias{viz_server}
\title{Visualization Module Server}
\usage{
viz_server(
  id,
  standard_data = cpiaetl::standard_cpia,
  africaii_data = cpiaetl::africaii_cpia,
  group_standard_data = cpiaetl::group_standard_cpia,
  group_africaii_data = cpiaetl::group_africaii_cpia
)

viz_server(
  id,
  standard_data = cpiaetl::standard_cpia,
  africaii_data = cpiaetl::africaii_cpia,
  group_standard_data = cpiaetl::group_standard_cpia,
  group_africaii_data = cpiaetl::group_africaii_cpia
)
}
\arguments{
\item{id}{Module namespace ID}

\item{standard_data}{Standard CPIA dataset (default: cpiaetl::standard_cpia)}

\item{africaii_data}{African Integrity Indicators CPIA dataset (default: cpiaetl::africaii_cpia)}

\item{group_standard_data}{Standard group averages dataset (default: cpiaetl::group_standard_cpia)}

\item{group_africaii_data}{African Integrity Indicators group averages dataset (default: cpiaetl::group_africaii_cpia)}
}
\description{
Visualization Module Server

Visualization Module Server
}
\keyword{internal}
