% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module.R, R/viz_ui.R
\name{viz_ui}
\alias{viz_ui}
\title{Visualization Module UI}
\usage{
viz_ui(id)

viz_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Visualization Module UI

Visualization Module UI
}
\keyword{internal}
